+++ packages/core/src/components/textarea/textarea.interface.ts
@@ -1,0 +1,3 @@
+export interface TextAreaChangeEventDetail {
+  value: string | undefined | null;
+}